highlight: extract to file
